requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
pymilvus==2.5.0
sentence-transformers==3.3.1
transformers==4.44.0
torch==2.5.0
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
numpy==1.26.4
streamlit==1.41.1
requests==2.32.3
python-dotenv==1.0.1